+++ source/utils.hpp
@@ -36,7 +36,7 @@ using Vec2d = Vec2<double>;
 using Vec2i = Vec2<int>;
 using Vec2u = Vec2<unsigned int>;
 
-void sleepThread(unsigned int seconds);
+void sleepThread(float seconds);
 void vBlank();
